Manifest update: Rust 'aux01' branch source

+++ rust/source/aux01/Assembly-CSharp/LensDistortion.cs
@@ -48,7 +48,7 @@ public sealed class LensDistortion : PostProcessEffectSettings
 
 	public override bool IsEnabledAndSupported (PostProcessRenderContext context)
 	{
-		if (enabled.value && !Mathf.Approximately ((float)intensity, 0f) && ((float)intensityX > 0f || (float)intensityY > 0f)) {
+		if (enabled.value && !Mathf.Approximately (intensity, 0f) && ((float)intensityX > 0f || (float)intensityY > 0f)) {
 			return !RuntimeUtilities.isVREnabled;
 		}
 		return false;
